Modify wrapuptime column for queue table
git-svn-id: https://rcs.lan.proformatique.com/svn/xivo/trunk@1128 dc85baef-7018-0410-bfbd-8a639da648bb

+++ base-config/asterisk/astmysql.db.sql
@@ -151,7 +151,7 @@ CREATE TABLE `queue` (
  `announce-round-seconds` tinyint unsigned,
  `announce-holdtime` varchar(4),
  `retry` tinyint unsigned,
- `wrapuptime` tinyint unsigned,
+ `wrapuptime` int(11) unsigned,
  `maxlen` int(11) unsigned,
  `servicelevel` int(11),
  `strategy` varchar(11),
